excel 양식 수정 및 BOOKING CONFIRMATION 다운로드만 양식 적용

+++ intranet/src/main/java/com/intranet/kch/model/entity/IVExcelEntity.java
@@ -21,5 +21,5 @@ public class IVExcelEntity extends Base {
     private LocalDate invoiceDate;
     private LocalDateTime startDate;
     private LocalDateTime endDate;
-    private Long totalPrice;
+    private Double totalPrice;
 }
